travis: initial setup automated testing on travis-ci
.travis.yml is used for automatic builds on travis build farm
(https://travis-ci.org/) if the travis service hook is enabled
for the repo on github.

This initial setup runs a whole build matrix for OSX and Linux,
using several python and java versions. The logs look like this:
https://travis-ci.org/rudimeier/cbstools-public

Note our build and test script ".travis-functions.sh" looks a bit
complicated because we don't have a real build system ...

BTW the java build is still not complete since I still don't know
what are the right mipav and plugin dependencies.